Created a helper function for category delete api
